Trying to fix failing tests. Not finished yet.
The main reason I'm committing at this point is to keep this state of the code
because it's acting so strangely. I have a failing test that indicates two
variables need to be swapped. If I do, the test tells my to swap them back.

This happens in perl 5.14.3. It seems to be different at least in later
versions, although 5.20.0 has more failing tests.